separate source to headers
